Move unit tests under test/unit
Doing this in prep of adding a second teenytest suite under test/safe
