+++ test/unit/function/index.test.js
@@ -1,15 +1,15 @@
-import Double from '../../src/value/double'
+import Double from '../../../src/value/double'
 
     // 3 throws it in the "store"
     //   - assigns the name to the entry in the store (if it exists)
 let create, imitate, remember, subject
 module.exports = {
   beforeEach: () => {
-    create = td.replace('../../src/function/create').default
-    imitate = td.replace('../../src/imitate').default
-    remember = td.replace('../../src/function/remember').default
+    create = td.replace('../../../src/function/create').default
+    imitate = td.replace('../../../src/imitate').default
+    remember = td.replace('../../../src/function/remember').default
 
-    subject = require('../../src/function/index').default
+    subject = require('../../../src/function/index').default
   },
   'pass in a name': () => {
     const double = new Double(null, null, 'fake thing')
